# [time] [coqc] Print time for document setup and end.