--- a/doc/changelog/09-cli-tools/17150-time_require_import.rst
+++ b/doc/changelog/09-cli-tools/17150-time_require_import.rst
@@ -1,0 +1,5 @@
+- **Added:**
+  ``coqc -time`` will now also report times of document creation and
+  ``.vo`` saving.
+  (`#17150 <https://github.com/coq/coq/pull/17150>`_,
+  by Emilio Jesus Gallego Arias).