--- a/test-suite/coq-makefile/timing/per-file-after/A.v.timing.diff.desired
+++ b/test-suite/coq-makefile/timing/per-file-after/A.v.timing.diff.desired
@@ -5,7 +5,7 @@
 0m14.578s | Chars 163 - 208 [Definition~foo1~:=~Eval~comp~i...] | 0m00.356s || +0m14.22s | +3994.94%
 0m00.284s | Chars 069 - 162 [Definition~foo0~:=~Eval~comp~i...] | 0m00.225s || +0m00.05s |   +26.22%
 0m00.140s | Chars 000 - 026 [Require~Coq.ZArith.BinInt.]        | 0m00.139s || +0m00.00s |    +0.71%
-0m00.024s | Chars 000 - 000 [Document~Start]        | 0m00.026s || -0m00.00s |   -7.69%
-0m00.004s | Chars 209 - 210 [Document~End]          | 0m00.004s || +0m00.00s |   +0.00%
+0m00.024s | Chars 000 - 000 [Comments~"Document~Start"]         | 0m00.026s || -0m00.00s |   -7.69%
+0m00.004s | Chars 209 - 210 [Comments~"Document~End"]           | 0m00.004s || +0m00.00s |   +0.00%
    0m00.s | Chars 027 - 068 [Declare~Reduction~comp~:=~nati...] |    N/A    || +0m00.00s |    N/A   
    N/A    | Chars 027 - 068 [Declare~Reduction~comp~:=~vm_c...] |    0m00.s || +0m00.00s |    N/A