--- a/test-suite/output-modulo-time/qed_time.out
+++ b/test-suite/output-modulo-time/qed_time.out
@@ -1,5 +1,8 @@
+Chars 0 - 0 [Comments~"Document~Start".]secs (u,s)
 Chars 40 - 57 [Lemma~foo~:~True.] 0. secs (0.u,0.s)
 Chars 58 - 64 [Proof.] 0. secs (0.u,0.s)
 Chars 67 - 81 [Axiom~(X~:~nat).] 0. secs (0.u,0.s)
 Chars 84 - 92 [exact~I.] 0. secs (0.u,0.s)
 Chars 93 - 97 [Qed.] 0. secs (0.u,0.s)
+Chars 98 - 99 [Comments~"Document~End".]secs (u,s)
+