--- a/toplevel/coqrc.ml
+++ b/toplevel/coqrc.ml
@@ -21,7 +21,7 @@ let load_rcfile ~rcfile ~state =
       match rcfile with
       | Some rcfile ->
         if CUnix.file_readable_p rcfile then
-          Vernac.load_vernac ~echo:false ~interactive:false ~check:true ~state rcfile
+          Vernac.load_vernac ~echo:false ~interactive:false ~check:true ~state rcfile |> fst
         else raise (Sys_error ("Cannot read rcfile: "^ rcfile))
       | None ->
         try
@@ -32,7 +32,7 @@ let load_rcfile ~rcfile ~state =
             Envars.home ~warn / "."^rcdefaultname^"."^Coq_config.version;
             Envars.home ~warn / "."^rcdefaultname
           ] in
-          Vernac.load_vernac ~echo:false ~interactive:false ~check:true ~state inferedrc
+          Vernac.load_vernac ~echo:false ~interactive:false ~check:true ~state inferedrc |> fst
         with Not_found -> state
         (*
         Flags.if_verbose