--- a/vernac/topfmt.ml
+++ b/vernac/topfmt.ml
@@ -428,7 +428,7 @@ let with_output_to_file fname func input =
 (* For coqtop -time, we display the position in the file,
    and a glimpse of the executed command *)
 
-let pr_cmd_header com =
+let pr_header ?loc com =
   let shorten s =
     if Unicode.utf8_length s > 33 then (Unicode.utf8_sub s 0 30) ^ "..." else s
   in
@@ -438,10 +438,11 @@ let pr_cmd_header com =
         | x -> x
       ) s
   in
-  let (start,stop) = Option.cata Loc.unloc (0,0) com.CAst.loc in
-  let safe_pr_vernac x =
-    try Ppvernac.pr_vernac x
-    with e -> str (Printexc.to_string e) in
-  let cmd = noblank (shorten (string_of_ppcmds (safe_pr_vernac com)))
+  let (start,stop) = Option.cata Loc.unloc (0,0) loc in
+  let cmd = noblank (shorten (string_of_ppcmds com))
   in str "Chars " ++ int start ++ str " - " ++ int stop ++
      str " [" ++ str cmd ++ str "] "
+
+let pr_cmd_header com =
+  let compp = try Ppvernac.pr_vernac com with e -> str (Printexc.to_string e) in
+  pr_header ?loc:com.CAst.loc compp