--- a/vernac/topfmt.mli
+++ b/vernac/topfmt.mli
@@ -72,4 +72,5 @@ val print_err_exn : exn -> unit
     redirected to a file [file] *)
 val with_output_to_file : string -> ('a -> 'b) -> 'a -> 'b
 
+val pr_header : ?loc:Loc.t -> Pp.t -> Pp.t
 val pr_cmd_header : Vernacexpr.vernac_control -> Pp.t