This can be useful when benchmarking.

As header in the time output, we use a special marker with a Coq
comment so it is still valid .v syntax: "(* Document Start / End *)"