@ejgallego 2023-01-23
> Timing the end of the document could also be nice, do you think you could include it in this PR?

Yes, that's a good idea. Do you mainly mean `save_library_to` + (Stm.join) ?