@ejgallego 2023-01-25
> Test suite broken

Fixed, note that I've updated the PR as not to produces a spurious "while in the prelude" line which is IMHO useless, moreover such stuff can be debugged just using `-noinit` + `From Coq Require Import Prelude`