@ejgallego 2023-01-26
Yes @gares , good idea.

@coqbot bench now

@SkySkimmer , the osx test suite stuff is fishy.