@SkySkimmer 2023-01-29
For https://coq.gitlab.io/-/coq/-/jobs/3667396595/artifacts/_bench/html/coq-fiat-core/src/ADT.v.html we have
- OLD timing
~~~
Chars 0 - 45 [Require~Export~Fiat.ADT.ADTSig...] 0.168 secs (0.141u,0.027s)
~~~
- NEW timing
~~~
Chars 0 - 0 [Document~Start] 0.025 secs (0.019u,0.006s)
Chars 0 - 45 [Require~Export~Fiat.ADT.ADTSig...] 0.176 secs (0.152u,0.023s)
Chars 46 - 47 [Document~End] 0.034 secs (0.03u,0.003s)
~~~
- HTML
~~~html
<div class="code" title="File: ADT.v
Line: 1

Time1: 0.168s
Time2: 0.025s
">
<div class="time1" style="width: 95.454545454545%"></div>
<div class="time2" style="width: 14.204545454545%"></div>
<pre>Require Export Fiat.ADT.ADTSig Fiat.ADT.Core.</pre>
</div>
<div class="code" title="File: ADT.v
Line: 2

Time1: 0s
Time2: 0.176s
">
<div class="time1" style="width: 0%"></div>
<div class="time2" style="width: 100%"></div>
<pre>

</pre>
</div>
~~~

It got confused by the non matching outputs, we should run the bench with the new printing on both commits.