@gares 2023-02-08
This is good for testing future PR against master. But it will make it impossible to compare, say, Coq 8.17 with 8.18.
Was the error message understandable? If not it should be fixed.