@gares 2023-02-08
Well, if it is not too hard to implement I'd prefer a message saying the two commits are not comparable because of this pr.