@ejgallego 2023-02-08
Wouldn't it be possible for the function doing the diff to detect this case and have it working by just removing the 2 extra additions if the other diff doesn't have them?