@ejgallego 2023-07-19
> What was left to be done?

@JasonGross I think the last bit to do here was to fix the timing comparison functions so things work or bail out without producing trash when we have a timing made with a version that outputs the end / start timing data.

In a sense, this is related to the problem of comparing two timing datasets of the same document but a different versions. I'm quite interested in this for `coq-lsp`, as I'd like to add to the performance browser a setting to display performance differences among different `git` revisions of a given .v file.